+++ src/world/astronomy/starfield.rs
@@ -6,6 +6,9 @@ use bevy::{
     },
 };
 
+#[derive(Component)]
+pub struct Parallax(pub f32);
+
 /// The setup function
 pub(crate) fn spawn_starfield(mut commands: Commands, assets: Res<AssetServer>) {
     let sampler_desc = ImageSamplerDescriptor {
@@ -38,5 +41,19 @@ pub(crate) fn spawn_starfield(mut commands: Commands, assets: Res<AssetServer>) 
         },
         NoFrustumCulling,
         Name::new("Background"),
+        Parallax(0.5),
     ));
+}
+
+pub(crate) fn parallax_effect(
+    mut parallaxes: Query<(&mut Transform, &Parallax)>,
+    camera: Query<&Transform, (With<Camera>, Without<Parallax>)>,
+) {
+    let camera_transform = camera.single();
+    let camera_translation = camera_transform.translation.truncate();
+
+    for (mut transform, parallax) in parallaxes.iter_mut() {
+        let z = transform.translation.z;
+        transform.translation = (camera_translation * parallax.0).extend(z);
+    }
 }